+++ mlprogram/actions/action_sequence.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 from mlprogram.actions.action \
     import Action, ApplyRule, Rule, ExpandTreeRule, NodeConstraint, NodeType, \
     GenerateToken, CloseVariadicFieldRule
-from mlprogram.asts import AST, Node, Leaf, Field, Root
+from mlprogram.languages.ast import AST, Node, Leaf, Field, Root
 from mlprogram import logging
