+++ mlprogram/languages/csg/functions.py
@@ -7,7 +7,7 @@ from mlprogram.languages.csg import Dataset
 from mlprogram.encoders import Samples
 from mlprogram.actions \
     import ActionSequence, ApplyRule, CloseVariadicFieldRule, Rule
-from mlprogram.asts import AST, Node, Field, Leaf, Root
+from mlprogram.languages.ast import AST, Node, Field, Leaf, Root
 from mlprogram import logging
 
 logger = logging.Logger(__name__)
