+++ test/languages/csg/test_functions.py
@@ -1,6 +1,6 @@
 import unittest
 from mlprogram.interpreters import Reference as R
-from mlprogram.asts import Node, Leaf, Field
+from mlprogram.languages.ast import Node, Leaf, Field
 from mlprogram.actions import AstToActionSequence
 from mlprogram.languages.csg import GetTokenType, IsSubtype, ToAst, ToCsgAst
 from mlprogram.languages.csg import Circle, Rectangle, Reference
