+++ test/languages/python/test_ast_to_python_ast.py
@@ -1,7 +1,7 @@
 import unittest
 import ast as python_ast
 
-import mlprogram.asts as ast
+import mlprogram.languages.ast as ast
 
 from mlprogram.languages.python.ast_to_python_ast import to_builtin_type
 from mlprogram.languages.python.ast_to_python_ast import to_python_ast
