+++ test/languages/python/test_python_ast_to_ast.py
@@ -1,7 +1,7 @@
 import unittest
 import ast as python_ast
 
-import mlprogram.asts as ast
+import mlprogram.languages.ast as ast
 
 from mlprogram.languages.python.python_ast_to_ast import base_ast_type
 from mlprogram.languages.python.python_ast_to_ast import to_ast
